adding lastCFMMFeeIndex as storage variable

+++ contracts/interfaces/strategies/events/IStrategyEvents.sol
@@ -28,6 +28,6 @@ interface IStrategyEvents {
     /// @param lpInvariant - lpTokenBalance as invariant units
     /// @param borrowedInvariant - lpTokenBorrowedPlusInterest as invariant units
     /// @param txType - transaction type. Possible values come from enum TX_TYPE
-    event PoolUpdated(uint256 lpTokenBalance, uint256 lpTokenBorrowed, uint96 lastBlockNumber, uint96 accFeeIndex,
+    event PoolUpdated(uint256 lpTokenBalance, uint256 lpTokenBorrowed, uint48 lastBlockNumber, uint96 accFeeIndex,
         uint256 lpTokenBorrowedPlusInterest, uint128 lpInvariant, uint128 borrowedInvariant, TX_TYPE indexed txType);
 }
